requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
psycopg2>=2.8.6,<2.8.7
django-bootstrap-form>=3.4    
django-cors-headers>=3.7.0  
django-filter>=2.4.0  
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.6.0
postgis>=1.0.4
djangorestframework-gis>=0.17
django-currentuser>=0.5.2